dts/freescale/meticulous: fix RTC communication on Var-Som Nano with Fika module
The Var-Som Nano was not communicating with the RTC module on
the Fika board due to misconfiguration in the device tree source.
This commit relocates the RTC module configuration from an
incorrect node to the correct I2C node, ensuring proper
initialization and communication.

The RTC device used is the DS1337

Signed-off-by: Mimoja <git@mimoja.de>